refactor(request-with-cache): move `calculateCacheKey` to common utils
This helper may be reused to calculate a cache key for another request
wrapper
